Complete core requirements of the project
Implemented the functionality of the frontend using react. Added 3 features:displaying the campaigns with their id/name and their play counts, auto refreshing every 3 seconds, using useState and setInterval, and a button which simulates an event POST that is sent to the database, processed and displayed -> show casing the whole data flow and the async processing of the task itself.